+++ src/main/java/com/qa/UserApiApplication.java
@@ -1,4 +1,4 @@
-package com.qa.userApi;
+package com.qa;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
