+++ src/main/java/com/qa/business/service/Users/UsersService.java
@@ -1,5 +1,7 @@
 package com.qa.business.service.Users;
 
+import java.util.List;
+
 import org.springframework.http.ResponseEntity;
 
 import com.qa.repository.domain.Users;
@@ -11,4 +13,6 @@ public interface UsersService {
     ResponseEntity<Object> deleteUser(String username);
 
     ResponseEntity<Object> updateUser(Users user);
+    
+    List<Users> getAll();
 }
