--- src/main/java/com/qa/repository/domain/UserRoles.java
@@ -1,52 +1,0 @@
-package com.qa.repository.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import javax.validation.constraints.NotBlank;
-
-
-@Entity
-@Table(name = "user_roles")
-public class UserRoles {
-	@Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long user_role_id;
-
-    @NotBlank
-    @Column(unique=true)
-    private String username;
-
-    @NotBlank
-    private String role;
-
-	public Long getUser_role_id() {
-		return user_role_id;
-	}
-
-	public void setUser_role_id(Long user_role_id) {
-		this.user_role_id = user_role_id;
-	}
-
-	public String getUsername() {
-		return username;
-	}
-
-	public void setUsername(String username) {
-		this.username = username;
-	}
-
-	public String getRole() {
-		return role;
-	}
-
-	public void setRole(String role) {
-		this.role = role;
-	}
-    
-    
-}
