+++ src/main/java/com/qa/rest/UsersRest.java
@@ -1,17 +1,19 @@
 package com.qa.rest;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.qa.business.service.UserRoles.UserRolesService;
 import com.qa.business.service.Users.UsersService;
 import com.qa.repository.domain.Users;
 
@@ -22,10 +24,7 @@ public class UsersRest {
 
     @Autowired
     private UsersService usersService;
-    
-    @Autowired
-    private UserRolesService userRolesService ;
-    
+
     
 //    @Value("${url.generator}")
 //    private String generatorURL;
@@ -35,20 +34,24 @@ public class UsersRest {
 //    
     
     
-    @DeleteMapping("/addUser")
+    @DeleteMapping("/deleteUser/{username}")
     public ResponseEntity<Object> deleteUser(@PathVariable String username) {
-        userRolesService.deleteUserRoles(username);
     	return usersService.deleteUser(username);
     }
 
-    @PutMapping("${path.updateBaby}")
+    @PutMapping("/updateUser")
     public ResponseEntity<Object> updateUser(@RequestBody Users user) {
         return usersService.updateUser(user);
     }
 
-    @GetMapping("${path.createBaby}")
+    @PostMapping("/createUser")
     public String createUser(@RequestBody Users user) {
         return usersService.addUser(user);
+    }
+
+    @GetMapping("/get")
+    public List<Users> getUser() {
+        return usersService.getAll();
     }
 
 }
