+++ src/main/resources/application.properties
@@ -1,6 +1,7 @@
+server.port=8081
 spring.datasource.url = jdbc:mysql://localhost:3306/userbd?useSSL=false
-spring.datasource.username = root
-spring.datasource.password = password
+spring.datasource.username =root
+spring.datasource.password =password
 
 
 ## Hibernate Properties
